Adds missing LICENSE file
The MIT license imposes to reproduce the MIT license, and
the xstatic-angular-fileupload package didn't do that.
This patch adds the missing LICENSE file, so that we
aren't doing anything illegal anymore.

Also bumps release in xstatic/pkg/angular_fileupload/__init__.py
so we can re-release the package in PyPi.

This patch also adds the missing gitreview file.

Change-Id: Ia92e2638392f3750e0bbd6b83ec3bc8a28dfd343